+++ refena/lib/refena.dart
@@ -17,6 +17,7 @@ export 'package:refena/src/notifier/base_notifier.dart'
         NotifyStrategy,
         RebuildableNotifier,
         FutureProviderNotifier,
+        StreamProviderNotifier,
         ViewProviderNotifier,
         ViewFamilyProviderNotifier,
         ReduxNotifier,
@@ -36,7 +37,6 @@ export 'package:refena/src/notifier/types/immutable_notifier.dart';
 export 'package:refena/src/notifier/types/notifier.dart';
 export 'package:refena/src/notifier/types/pure_notifier.dart';
 export 'package:refena/src/notifier/types/state_notifier.dart';
-export 'package:refena/src/notifier/types/stream_provider_notifier.dart';
 export 'package:refena/src/observer/error_parser.dart' show ErrorParser;
 export 'package:refena/src/observer/event.dart';
 export 'package:refena/src/observer/history_observer.dart';