+++ refena/lib/src/proxy_ref.dart
@@ -86,7 +86,7 @@ class ProxyRef implements Ref {
   }
 
   @override
-  Future<T> future<N extends BaseAsyncNotifier<T>, T>(
+  Future<T> future<N extends GetFutureNotifier<T>, T>(
     BaseProvider<N, AsyncValue<T>> provider,
   ) {
     if (_onAccessNotifier == null) {